backend/requirements.txt
asgiref
Django
django-jazzmin 
pip
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
tzdata 

 